+++ database/migrations/2025_06_27_104435_create_sp_get_food_package_details_by_id.php
@@ -10,7 +10,11 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up(): void
-    {
+    {       
+        if (config('database.default') !== 'mysql') {
+            return;
+        }
+
         DB::unprepared('DROP PROCEDURE IF EXISTS sp_get_food_package_details_by_id');
 
         DB::unprepared('
@@ -67,7 +71,10 @@ return new class extends Migration
      * Reverse the migrations.
      */
     public function down(): void
-    {
+    {   
+        if (config('database.default') !== 'mysql') {
+            return;
+        }
         Schema::dropIfExists('DROP PROCEDURE IF EXISTS sp_get_food_package_details_by_id');
     }
 };
